修改自定义协议头；可以根据用户id、别名、标签推送；支持离线消息接收

+++ app/src/main/java/com/mpush/android/MPushService.java
@@ -120,7 +120,7 @@ public final class MPushService extends Service implements ClientListener {
     }
 
     @Override
-    public void onReceivePush(Client client, byte[] content, int messageId) {
+    public void onReceivePush(Client client, byte[] content, long messageId) {
         sendBroadcast(new Intent(ACTION_MESSAGE_RECEIVED)
                 .addCategory(BuildConfig.APPLICATION_ID)
                 .putExtra(EXTRA_PUSH_MESSAGE, content)
